Wired up kafkaSource with KafkaSourceEpochState
Cleaned up some code

+++ src/main/scala/org/codefeedr/core/library/internal/kafka/meta/SourceEpoch.scala
@@ -3,4 +3,4 @@ package org.codefeedr.core.library.internal.kafka.meta
 import org.codefeedr.model.zookeeper.Partition
 
 //Description of a source Epoch
-case class SourceEpoch(partitions: List[Partition], epochId: Int, subjectEpochId: Int)
+case class SourceEpoch(partitions: List[Partition], epochId: Long, subjectEpochId: Long)
